+++ apps/blog/src/components/arrow-link.astro
@@ -1,7 +1,7 @@
 ---
 import Link from "@/components/site-nav-item.astro";
 import { FaArrowLeftLong, FaArrowRightLong } from "react-icons/fa6";
-import { cn } from "@/utils";
+import { cn } from "@/lib/utils";
 import type { ComponentProps } from "astro/types";
 
 const { label, ...props } = Astro.props;
