+++ apps/blog/src/components/article-header.astro
@@ -2,8 +2,9 @@
 import Link from "@/components/ui/link.astro";
 import { format } from "date-fns";
 import { LuCalendarDays, LuTags, LuDot, LuCoffee } from "react-icons/lu";
-import { cn, slug, type Entry } from "@/utils";
+import { cn, slug } from "@/lib/utils";
 import type { HTMLAttributes } from "astro/types";
+import type { Entry } from "@/lib/content";
 
 const { class: className, data, ...props } = Astro.props;
