+++ apps/blog/src/components/mdx/hr.astro
@@ -1,6 +1,6 @@
 ---
 import { Separator } from "@/components/ui/separator";
-import { cn } from "@/utils";
+import { cn } from "@/lib/utils";
 import type { SeparatorProps } from "@radix-ui/react-separator";
 import type { HTMLAttributes } from "astro/types";
