+++ apps/blog/src/components/site-footer.astro
@@ -1,6 +1,6 @@
 ---
 import Link from "@/components/ui/link.astro";
-import { cn } from "@/utils";
+import { cn } from "@/lib/utils";
 import type { HTMLAttributes } from "astro/types";
 
 const { class: className } = Astro.props;
