+++ apps/blog/src/components/ui/button.astro
@@ -1,6 +1,6 @@
 ---
 import { cva, type VariantProps } from "class-variance-authority";
-import { cn } from "@/utils";
+import { cn } from "@/lib/utils";
 import type { HTMLTag, Polymorphic } from "astro/types";
 
 type Props<Tag extends HTMLTag> = Polymorphic<{ as: Tag }> &
