+++ apps/blog/src/components/ui/icon.astro
@@ -1,5 +1,5 @@
 ---
-import { cn, slug } from "@/utils";
+import { cn, slug } from "@/lib/utils";
 
 import { LuHash } from "react-icons/lu";
 import * as si from "react-icons/si";
