+++ apps/blog/src/layouts/base-layout.astro
@@ -3,7 +3,7 @@ import "@/styles/globals.css";
 
 import VercelInsights from "@/components/vercel-insights.astro";
 import { siteConfig } from "@/config";
-import { cn } from "@/utils";
+import { cn } from "@/lib/utils";
 import type { HTMLAttributes } from "astro/types";
 
 const { head, class: className, ...props } = Astro.props;
