+++ apps/blog/src/pages/index.astro
@@ -1,7 +1,7 @@
 ---
 import Layout from "@/layouts/home-layout.astro";
 import PostIndex from "@/components/post-index.astro";
-import { getEntries } from "@/utils";
+import { getEntries } from "@/lib/content";
 import { AiFillHeart } from "react-icons/ai";
 import { LuRss } from "react-icons/lu";
